+++ src/emmental/metrics/roc_auc.py
@@ -1,6 +1,10 @@
+import logging
+
 from sklearn.metrics import roc_auc_score
 
 from emmental.utils.utils import pred_to_prob
+
+logger = logging.getLogger(__name__)
 
 
 def roc_auc_scorer(golds, probs, preds, uids=None, pos_label=1):
@@ -19,6 +23,14 @@ def roc_auc_scorer(golds, probs, preds, uids=None, pos_label=1):
     """
 
     gold_probs = pred_to_prob(golds, n_classes=probs.shape[1])
-    roc_auc = roc_auc_score(gold_probs, probs)
+
+    try:
+        roc_auc = roc_auc_score(gold_probs, probs)
+    except ValueError:
+        logger.warning(
+            "Only one class present in golds."
+            "ROC AUC score is not defined in that case, set as nan instead."
+        )
+        roc_auc = float("nan")
 
     return {"roc_auc": roc_auc}